requirements.txt
Django==3.2.18
environs==9.5.0
python-telegram-bot==13.13
psycopg2-binary==2.9.7
gspread==5.11.3